+++ driver/sql/postgres/internal/bigint/doc.go
@@ -1,0 +1,3 @@
+// Package unsigned provides a type maps an unsigned 64-bit integer to a signed
+// 64-bit integer for use in SQL statements and scan operations.
+package bigint