+++ CHANGELOG.md
@@ -10,6 +10,16 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+### Unsigned
+
+### Changed
+
+- **[BC]** The `pgjournal` implementation now supports the full range of
+  unsigned `journal.Position` values. The `uint64` value-space is "mapped" onto
+  the `int64` space such that `uint64(0) == math.MinInt64`, thereby preserving
+  order. The `journal.position` column has been renamed to `encoded_position` to
+  make it clearer that the value can not be used as-is.
+
 ## [0.6.0] - 2024-03-16
 
 ### Added